requirements.txt
Flask
google-cloud-speech
google-cloud-texttospeech
genai
